requirements.txt
backcall==0.2.0
certifi==2021.10.8
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
fonttools==4.29.1
imageio==2.14.1
ipykernel==6.8.0
ipython==7.31.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.26
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pywin32==303
pyzmq==22.3.0
six==1.16.0
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
wincertstore==0.2
